Update transaction details with correct transfer data

+++ service/src/main/kotlin/io/provenance/explorer/domain/ExplorerModels.kt
@@ -17,7 +17,7 @@ data class BlockDetail(val height: Int, val time: String, val proposerAddress: S
                        val votingPower: Int, val numValidators: Int, val txNum: Int, val bondedTokenPercent: Int, val bondedTokenAmount: Int,
                        val bondedTokenTotal: Int)
 
-data class TxDetails(val height: Int, val gasUsed: Int, val gasWanted: Int, val gasLimit: Int, val gasPrice: Int, val time: String,
+data class TxDetails(val height: Int, val gasUsed: Int, val gasWanted: Int, val gasLimit: Int, val gasPrice: BigDecimal, val time: String,
                      val status: String, val fee: BigDecimal, val feeDenomination: String, val signer: String,
                      val memo: String, val txType: String, val from: String, val amount: Int, val denomination: String, val to: String)
